refactor(WorldClock): restructure WorldClock component by removing old implementation and updating index file for cleaner exports
